Refactoring Domain configuration preparation
